backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
python-multipart
replicate
google-cloud-firestore
google-genai>=1.52.0
